sys_src/backend/requirements.txt
anyio==3.6.2
fastapi==0.95.1
httpx==0.23.1
h11==0.14.0
sniffio==1.3.0
starlette==0.26.1
pydantic == 1.10.7
pymongo == 4.3.3
uvicorn==0.22.0
kafka-python==2.0.2
websockets == 11.0.3
wsproto == 1.2.0